+++ src/FinServ.Application/Handlers/Produtos/QueryAvailableProdutos/QueryAvailableProdutosResponse.cs
@@ -4,7 +4,7 @@
     {
         public int IdProduto { get; set; }
         public string Nome { get; set; }
-        public double TaxaJurosMensal { get; set; }
+        public decimal TaxaJurosMensal { get; set; }
         public string DataVencimento { get; set; }
         public int Quantidade { get; set; }
         public int CodigoProduto { get; set; }
