+++ src/FinServ.Domain/Entities/Ativos/ExtratoAtivos.cs
@@ -10,12 +10,12 @@ namespace FinServ.Domain.Entities.Ativos
         public string NomeProduto { get; set; }
         public string DescricaoProduto { get; set; }
         public DateTime DataVencimentoProduto { get; set; }
-        public double TaxaJurosMes { get; set; }
-        public double ValorCompra { get; set; }
-        public double ValorAtual { get; set; }
+        public decimal TaxaJurosMes { get; set; }
+        public decimal ValorCompra { get; set; }
+        public decimal ValorAtual { get; set; }
         public int Quantidade { get; set; }
         public DateTime DataCompra { get; set; }
-        public double Rentabilidade { get; set; }
-        public double ValorPatrimonio { get; set; }
+        public decimal Rentabilidade { get; set; }
+        public decimal ValorPatrimonio { get; set; }
     }
 }
