+++ src/FinServ.Domain/Entities/Produtos/Produto.cs
@@ -9,11 +9,11 @@ namespace FinServ.Domain.Entities.Produtos
         public string Nome { get; set; }
         public string Descricao { get; set; }
         public int CodigoProduto { get; set; }
-        public double TaxaJurosMensal { get; set; }
+        public decimal TaxaJurosMensal { get; set; }
         public DateTime DataVencimento { get; set; }
         public int Quantidade { get; set; }
         public ICollection<Ativo> Ativos { get; set; }
-        public double Valor { get; set; }
+        public decimal Valor { get; set; }
 
 
     }
