+++ src/FinServ.Domain/Interfaces/Entities/IExtratoAtivos.cs
@@ -10,14 +10,14 @@ namespace FinServ.Domain.Interfaces.Entities
         public string NomeProduto { get; set; }
         public string DescricaoProduto { get; set; }
         public DateTime DataVencimentoProduto { get; set; }
-        public double TaxaJurosMes { get; set; }
-        public double ValorCompra { get; set; }
-        public double ValorAtual { get; set; }
+        public decimal TaxaJurosMes { get; set; }
+        public decimal ValorCompra { get; set; }
+        public decimal ValorAtual { get; set; }
         public int Quantidade { get; set; }
         public DateTime DataCompra { get; set; }
         [NotMapped]
-        public double Rentabilidade { get; set; }
+        public decimal Rentabilidade { get; set; }
         [NotMapped]
-        public double ValorPatrimonio { get; set; }
+        public decimal ValorPatrimonio { get; set; }
     }
 }
